Refactor code

+++ JewelrySalesSystem.BAL/Models/VnPays/VnPayRequest.cs
@@ -20,7 +20,7 @@ namespace JewelrySalesSystem.BAL.Models.VnPays
             this.vnp_CurrCode = "VND";
             this.vnp_CreateDate = createDate.ToString("yyyyMMddHHmmss");
             this.vnp_TmnCode = tmnCode;
-            this.vnp_Amount = (int)amount * 100;
+            this.vnp_Amount = amount * 100;
             this.vnp_Command = "pay";
             this.vnp_OrderType = orderType;
             this.vnp_OrderInfo = orderInfo;
